WA-23 services created

+++ src/main/java/com/weathe_alligator/user_service/repo/LocationRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LocationRepo extends JpaRepository<Location, Float> {
+public interface LocationRepo extends JpaRepository<Location, Long> {
 }
